Framework for picking up seeds
New Seed script holds just a name (Of the actual Seed prefab), and a
Number of (If we wish for a single new seed to hold multiple seed
instances)

Otherwise, commented out code in the player raycast for when we have
reorganized into Resources to load Prefabs that way.
Specifically: The 'name' string in a 'New_Seed' must match the name of
the Prefab we're trying to load.