So I fixed the beanstalk the reference goes to the name of the object so so long as it's called beanstalk(clone) it's perfect. This is great cause there are no numerical incrimented names cause each new beanstalk is placed in a new plot of dirt so it has exactly that name :D the mesh is also conforming directly to the beanstalk so if you're having trouble climbing you may just be bumping your head against the top of the mesh.
With regards to camera the third person still works and switches at the press of a switch. I am running into slight problems with the clipping into walls but i've gotten it so soft clips don't matter however hard clips still can update with the camera on another side of the wall. It doesn't disconnect from the player though so it should be fine for the moment i've send an email to Alex so hopefully he can help.
The first person camera i just needed to increase the radius of the player rigidbody slightly so the camera never gets close enough to a surface on any side to clip through it.  I've been coding solidly since midnight though (currently 6:22) so i'm going to bed.